Fix: Protect against segfault while printing a time_t variable. Detected in NetBSD6 RC2(32-bit)
Cherry-picked from: ce80a070dfcbc5919f4940ec872248859fb618e8

Conflicts:

	src/logging.c

+++ src/logging.c
@@ -502,7 +502,7 @@ if ((fout = fopen(filename,"a")) == NULL)
    return;
    }
 
-fprintf(fout,"%ld,%ld: %s\n",CFSTARTTIME,now,s);
+fprintf(fout,"%jd,%jd: %s\n",(intmax_t) CFSTARTTIME,(intmax_t) now, s);
 fclose(fout);
 }
